UPD: Service separation by entities
