+++ bot-api/EducationalTeamsBotApi/EducationalTeamsBotApi/EducationalTeamsBotApi.Domain/GlobalSuppressions.cs
@@ -1,0 +1,8 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1201:Elements should appear in the correct order", Justification = "<En attente>", Scope = "member", Target = "~M:EducationalTeamsBotApi.Domain.Entities.CosmosReaction.#ctor(System.String,System.String,System.Int32)")]
